rename utils lib to rps_utils to fix name conflict
