+++ cime_config/MOM_RPS/FType_MOM_params.py
@@ -1,6 +1,6 @@
 import os
 from MOM_RPS import MOM_RPS
-from utils import get_str_type
+from rps_utils import get_str_type
 
 class FType_MOM_params(MOM_RPS):
     """ Encapsulates data and read/write methods for MOM6 case parameter files: MOM_input, user_nl.
